move SphereMan into interactables folder for better organization
